Integrate uiautomatorviewer into Eclipse
This CL integrates uiautomatorviewer into Eclipse. There are two
components to this:
 1. Take UI Hierarchy Dump from a device: This is added as an action
    to the device view in the DDMS perspective.
 2. Viewing the resulting UI hierarchy: This is added as an editor,
    associated with the .uix extension. The editor simply wraps a
    UiAutomatorView.

The UiAutomatorView is modified slightly to allow for opening just
the hierarchy file without the associated screenshot. This is to
enable opening the .uix file via standard Eclipse Open File Dialog.
When opened this way, the canvas displaying the screenshot is replaced
with a button that allows the user to specify the actual screenshot
to be displayed.

Change-Id: I0f31fe95b2728123ff6f0f748478c18c0c2f1b48