Make @TargetApi quickfix use VERSION_CODES instead of ints
The quickfix to insert a @TargetApi annotation for an API
violation will now insert

    import android.os.Build;    (if necessary)
    ...
    @TargetApi(Build.VERSION_CODES.GINGERBREAD)

etc. Other than the fact that we're using the above format
in our templates, this also has the advantage that code
completion and show documentation (F2) with the caret on
the field will show the compatibility information for
that release, which isn't the case for raw integers.

Change-Id: I233473bf6df706dc73bc692af7ee059287bc22f1