gltrace: Default to displaying texture mipmap level 0.
Textures can have multiple mipmap levels. To view the uploaded
texture requires that you explicitly select a particular mipmap
level. However, in many cases, there is only 1 level (level 0).
In such cases, rather than having to drill down the hierarchy
and then select mipmap 0 to view the texture, this CL allows
just selecting the texture to trigger a display of the zeroth
mipmap if it is available.

Change-Id: I9f7f8f10f99945377c6cb7dfab19db75cea383e0