ADT: Detect when SDK platforms/addons might have changed.
This adds a new functionality in SdkManager to keep
track of the existing platforms/addons folder and
detect later whether they might have changed.

The check is a quick sanity check done on the presence
of the target directory, its last-modified timestamp
as well as the one of the source.properties.
Whenever an SDK Manager instance updates an existing
target or add/removes a new one, a change will be detected.

Non-goal: this does not trigger when a user manually
modifies the content of a target (e.g. its data files).

A method is added in AdtPlugin to reload the SDK if the
above method detects the targets have potentially changed.
This is then used when there's a user interaction with
something that depends on the latest state of the SDK,
namely: opening XML files (layout, manifest, etc.) or
opening a wizard (AVD manager, new project, new template.)
These wizards already register for sdk/target change
listeners and are notified asynchronously if the SDK has
changed and is refreshed.

Change-Id: Ia343c26321c0cb39c28a7c5d570c94e543344401