junit: Rework support for parallel junit launch.
A previous commit (2c57cbea0d663b) introduced support for
running Junit tests concurrently on all connected devices.
It relied on changing the TestIdentifier to store device
information, and having a single listener that listens to
test events from all devices.

The change to TestIdentifier caused issues with some users
of ddmlib in cases where the device info is not available
all the time.

This patch reverts all the changes in ddmlib, and moves
the knowledge of what device tests are being run to the
listener layer in ADT. So now we have a per device test
listener that knows only about tests run on that device.

Change-Id: Iffedcb38cdf21e349fbe38ecf9a654060b469b04