SDK Manager: fix "null" error some people are getting.
Some users reported in some cases they see a line in
the sdk manager log that just says "null". Turns out
to be an NPE when trying to read the null InputStream
returned by UrlOpener in case of a failed URL fetch.

UrlOpener used to throw an IOException when a resource
couldn't be fetched, but then I added an alternate
download path (to handle file:// resources) and that
made it fail with a null input stream instead. So now
the code throws an exception like it used to before.

Change-Id: I92ef53992290c5e6fde8b0400274bbee822f2a61