SDK Manager: socket read timeout.
This adds a non-null default timeouts on the HttpClient
and UrlConnection instances. Most important is the
socket read timeout which seems to be infinite by default.
The default is at 1 minute for that one, with an option
to change it via an env variable. I might want to expose
it in the options dialog later (in another CL.)

SDK Bug/Request: 26382

Change-Id: I8b1da6505ea331d0520987ab8955f39cf123202b