SDK Manager: fix HTTP Auth cancel.
When the SDK Manager prompts the user for HTTP Auth
or NTLM auth, the API says the result should be null
if the user cancelled.
It wasn't, it was instead empty string fields and
the download would abort the a "failed: no content"
error in the log.
This fixes the UI to respect the convention.

It also makes sure that if the HttpClient call
fails due to the user cancelling, we don't try
to fallback on the java.UrlConnection method.

Change-Id: Ibf1807f4becd40dbced17f851b3fdc5b394788ea