Don't ask about Hirarchical Parent Activity in new projects
This changeset hides the "Hierarchical Parent" setting from
the activity panel in the New Project Wizard.

It also adds "Optional" as a text field hint (where SWT supports
it) on this field when it's present (and on any other text fields
whose tempalte metadata states that they allow empty strings).

Change-Id: Id3334357b62fdaedcc8eb2ac08dac9a1f005efc6