Fix class to source mapping for the super() keyword
The API checker needs to look up the corresponding source code source
range for a given bytecode method dispatch, and the line number
information found in the .class file is not accurate enough. Lint has
some heuristics to search for the corresponding source range, but this
did not work properly for cases where the method call was a parent
constructor, via the "super()" call.

It also fixes a bug in handling the default package.

Change-Id: I15bb07b6709ab9a53376398fe92715f6e6eba968