Preserve fatal severity for Lint XML reports
The Lint CLI does not distinguish between errors and fatal errors in
the error output. However, changing the severity was happening when
the issue was reported rather in the output reporter itself, which
meant that it accidentally changed the severity in the XML report,
where all metadata should be preserved.

Change-Id: Ib3ee9a20a7b1d2413ed85e1952f5f3d1783a8826