Add target SDK setting to the New Project wizard
Since the first page is getting too large, rather than add
one more, split some of the contents into a second page.
In addition, the new project dialog was missing a workset
chooser like the old wizard had, so add one in on the new
page now that we have plenty room for it.  Some other tweaks
as well based on input from Xav.

Change-Id: I4dd2334b2f5af623cba4bbb3a6d2d0b053d1763f