Also check SimpleDateFormat calls for explicit locale
This changeset expands on the lint check for calling String
convenience methods that are locale sensitive without an explicit
locale (issue 36822). It checks whether a SimpleDateFormat constructor
that uses the default locale is called, and if so suggests either
specifying a locale or using one of the pre-configured locale
sensitive instances of SimpleDateFormat.

Change-Id: Ic4fcb72c2ee6b31d1534b4d909493fef7dcff867