Close XML editors when files are deleted or projects are closed
The existing hooks for closing the editor when the underlying file is
deleted was not working correctly, and was inefficient (since each
editor added their own global resource listener, so every editor would
find out about every other editor's file changes).

Instead, this generalizes the single editor listener which was used to
initialize editor types such that it also listens for deletion, and
then finds any open editor mapped to that file.  It also hooks up to
the pre-close events for projects and closes all files related to the
project as well.

This will hopefully fix this issue as well:
20836: Can't delete a layout XML via Package Explorer > Delete if it
       has an error

Finally, it removes a bunch of now obsolete logging code for an issue
which seems to be permanently fixed.

Change-Id: I90b38984639a605755f2d67ca2413cc925f730d1