Avoid "Marker id: 123456789 not found" errors
We have some code which runs asynchronously to add or clean up
error markers. In some cases this code runs after a project has
been closed or deleted, which results in errors. Guard these
marker access points with an IResource.isAccessible() check
(which performs both exists(), and for projects, isOpen()).

Change-Id: Ie4884db13b1e00236c1c219d95c3544fcdee9610