Refactor SDK Manager into "core" vs "ui" packages.
This is just a refactor to move classes around.
No functionality is actually changed.

The sdkman2 package contains classes that are now
sorted between "ui" and "core" packages. The core
stuff has no UI dependency and can be used in
unit tests, whereas the ui stuff needs SWT somehow.

In a later CL some of the UI stuff will be changed
to make it easier to mock for testing.

Change-Id: I86606df7992de6ca6ae1df95f7b712cbba4fd3b6