Improve last editor mode handling
We recently added the ability for the IDE to remember whether you last
edited an XML file in text mode or in graphical mode (see issue
31340). However, this was tracked with a single boolean flag, which
meant that it was an "all or nothing" flag.

However, you may want to always edit string resource files with the
XML editor, but you want to always use the graphical editor for the
manifest file.

In this changeset, the state is kept in a bitmask instead, which
allows us to track the broad editor types separately. With this, new
editors are opened according to the last mode you used for that type
of editor. (Note that it tracks "categories" of editors, not
individual resource types, so for example the editor for color
resources and the editor for state list drawables share the same last
mode as the string resource editor. This is because these editors
share roughly the same graphical editing capabilities.)

(The CL also contains some minor code cleanup.)

Change-Id: I4624dffa2349230684c0558a33081adda8d799b2