gltrace: Do not save thumbnail images in memory
These were originally saved with the idea that if the trace file
itself was overwritten after it was parsed, we could display
the thumbnail image atleast since we don't have access to the full
image anymore.

However, this hasn't turned out to be a common case, but it just
uses up memory for large traces.

Change-Id: I1152e23f29563f47b6818b89c5a50bf5ce2a5084