Make resource chooser include resources from library projects
The resource chooser was only showing resources from the current
project. This changeset fixes this such that it will merge in
resources from library projects as well.

It also cleans up the ResourceChooser class such that it has a couple
of convenience factory methods, and simplified various call sites to
use the factory instead.

Finally, it similarly fixes the Go To Declaration hyperlink resolver
to also look at library projects when resolving source locations.

Change-Id: Id20d14e4d9c6d9133b827ef3a3854e47c65b73fb