Don't flag missing classes if project has not been built yet
The missing class detector in lint will warn if the manifest file
contains references to class files and lint does not see those class
files in the class output folder (or in any of the library jars).
This is useful for catching typos and deleted classes.

However, if you run lint on a project that has not yet been built,
this can give misleading or confusing errors, since lint decides that
a class exists by whether a .class file is actually there, not by
whether a source file exists which would result in a class file
getting produced with that name.

Change-Id: I3105157a37f575e00b550ba90d883f24c02c16a9