Improve the Device menu
First, this changeset improves the device labels in the device menu to
not only list the screensize and screen type (e.g. "FWVGA"), but the
screen resolution and density category as well.

Second, the device menu by default will show one submenu for every
device manufacturer, with the corresponding devices listed within each
of those sub menus. Out of the box the default file only lists Nexus
devices, and this ends up with a bit of a suboptimal menu.

This changeset handles this scenario better by inlining all the
generic devices together in the top menu, and sorting them in reverse
order. It also places the Nexus devices in this menu, ordered by
reverse release date.

Change-Id: I5015430e2dc48306e848ce7691810349415470e0