logcat: Support auto scroll lock.
Another go at adding support for automatically enabling/disabling
scroll lock. Past attempts have been brittle because of differences
in how Windows, Linux and Mac handle the scrollbar.

The key to getting this work consistently across platforms seems
to be in explicitly setting the range of values for the scrollbar.
Setting the maximum to be a high enough value relative to the
thumb allows accurate detection of even minor movements away from
the bottom of the scrollbar.

Change-Id: Ic3259cc4e2e8a20f3a87ce4bf234217cea792f88