Improve aapt error message
If anything goes wrong with aapt, we emit an error message that we
could not execute aapt and to check the install location.

However, if the file exists but isn't executable, or if a parsing
error happens in the new SymbolLoader/Writer, the error message is
misleading. This changeset makes the error message more specific if it
looks like the root cause isn't a missing aapt installation.

Change-Id: Ifb09a2dd736f6a93c0f167162826e1ad6e3d2a63