Add workaround for the Eclipse+Mountain Lion bug
On Mountain Lion, the property sheet's property editor inline text
editor is broken. This is due to
https://bugs.eclipse.org/bugs/show_bug.cgi?id=388574

The bug has been fixed for Eclipse 4.3, but right now it's a very poor
user experience; you click to edit, and nothing happens.

This changeset adds a workaround: *If* the bug is triggered (an
exception is thrown during property editor activation), then the
customizer dialog is shown instead (equivalent to clicking on the
"..." button next to the property).

The dialog code is modified to show an error message on the bottom
explaining why the dialog was shown instead (since this behavior is
annoying, we should explain to users why we are doing), and it
explains that the bug should be fixed in Eclipse 4.3, so if anyone in
the future is running Eclipse 4.2 with this behavior and 4.3 is
available, they will have an incentive to upgrade.

Change-Id: I338a26c39aca0042075d7a12527b723d9783509c