Include adapter binding support for included layouts
This changeset adds support for looking up adapter binding metadata
(such as which layout to use for a list view) for layouts that
included by other layouts (and potentially nested even deeper than
that).

This hasn't worked in the past because the adapter binding relied on
getting a UiViewElementNode, supplied as a view cookie by the element
parser. However, since we also provide the parser used for included
context (a simply KXml parser), it now looks up the tools attributes
and puts them in a map as a view cookie (for the tags known to have
adapter view bindings) and uses this later in the project callback to
look up a suitable adapter binding.

(Also tweak error message, and switch from rendered previews to
builtin icons for the imageview and imagebutton palette items.)

Change-Id: I99e8b10efb172a1918b74b11123d11c275e1f004