Synchronize layout metadata across layout configurations
If you choose a preview layout to use for a ListView or GridView, or
if you choose a fragment layout to use for a <layout> tag, until now
that setting has only applied to the exact layout you are currently
editing.

This changeset makes the setting also get synchronized to all the
other layout variations (-land, -xlarge etc) of the layout, such that
you don't have to find yourself tweaking this setting each time you
open some new variation of the layout after changing preview settings.

(Also deletes a bunch of the old comment-based metadata code.)

Change-Id: I712213920a8956901da157eb94611b2372368fcc