SDK fix for bug 21589: disable type rename refactoring.
As a simplified workaround for bug 21589, we disable the
Android Type Rename refactoring if the current rename
options enable the "update fully qualified names in non-java files"
and the file patterns mentions "xml".

It may generate false negatives (e.g. if the pattern doesn't
cover the AndroidManifest.xml or the layouts that would have been
renamed) but it seems safer than corrupting the XML file.

Change-Id: I5e7ba3c76c0e643225d03cf0016bf431497818e7