Add label to AVD Creation dialog for no emulator images installed
It's pretty easy to end up installing a recent Android platform
without also installing an emulator image such as ARM or x86.

When you create an AVD, and you select that platform, the OK button
is disabled, but it isn't obvious why; there's a Target: label
next to an empty and disabled combo box.

This CL simply sets the text of the combo box to "No system images
installed for this target." (while keeping the combobox disabled) to
hopefully guide the user in the right direction.

Change-Id: I5dcc3e75944440f04e1054be35fe68be41de734f