Multi-configuration editing bug fixes
This changeset fixes a number of bugs with multi configuration
editing; especially around configuration switching and inheritance, as
well as some memory improvements.

It also synchronizes values between layout variations. If you for
example have layout-land/foo.xml and layout-port/foo.xml and you
change the Theme to Theme.Holo or the device to Galaxy Nexus, the
configurations for both files will be updated (whether or not the
files are open), and provided of course the layout folder doesn't
contradict it; layout-xlarge/foo.xml would be unaffected by the above
edit since Galaxy Nexus doesn't match the layout-xlarge folder.

Change-Id: I5c01555aad8339f68788d8aed1f707d30993ae1b