Fix handling of @SuppressLint on fields
When you have a field initialization like this:

    private int foo = new ForbiddenClass();

the actual code to perform the initialization lives in a method named
<init> (or for a static field, <clinit>).

If you tried to suppress lint errors here by adding an annotation on
the field, it would not be found by lint, since lint looks at method
level annotations and it's sitting on the field node instead.

To fix this, the suppress check needs to identify the field scenario,
and in that case (when checking whether an error is suppressed) look
up the field in the class and check annotations there.

There was a second bug: The lint check which looks for invalid
suppress annotation sites (since they're not allowed on local
variables for class based lint checks) incorrectly concluded that
these types of initializations were local variables rather than field
initializations.

This fixes issue
38626: ADT: Misleading lint rule: SimpleDateFormat

Change-Id: I254f3fb5a6132d6cbe39bd425ffe6d67ed7b84ed