Automatically pick the best render target
This changeset adds a new option, on by default, which causes the
layout editor to always pick the best rendering target. The best
rendering target is typically the most recent one. This option can be
toggled in the rendering target drop down menu (but it applies
globally), and choosing a specific rendering target turns it off.

Change-Id: I53e48a38741f364f2b68525a4898aca69b1ae7b1