adtpackage: make adtpackage depend on, not include features
If adtpackage is declared to provide a bunch of features, then it
doesn't seem possible upgrade those features to a later version.

This CL makes adtpackage simply depend on those features being
present, which allows for upgrading those features at a later time.

In order to achieve this, we first build all the features, and
dump them in the repository. Then the product build simply takes
the collection of repositories, and the product definition, and
creates the product.

Change-Id: I8ff7a7ea05aea9f08d69c624499185ec1578fbdb