emma: XML and text reports as well
Since we have no way for a user to tell emma what kind of report we
want, just save all three types. The HTML one is the largest, the XML
report is usually less than 10% of the size, and the text report is
usually a few kilobytes.

Also fix the temporary file cleanup to delete the right coverage.em
file.

Change-Id: Ifc60fa4bf8a3b6b0e699e20dcbff8c818c19729b