Create new XML files in configuration of selection
If you invoke the New XML File Wizard with a selection
which implies a specific configuration (such as -v11
or -port), the New XML File Wizard will now initialize
the second page's folder configuration selection with
the same configuration. The net result is that right
clicking on values-v14 and selecting New XML File
will now create the file in that folder.

Change-Id: Ia874af628e8fa936700f58d6267b973414c24d6a