Additional fixes to filter out library project lint warnings
Lint already has the concept of whether a project should report lint
warnings or not; a project can be included in analysis (e.g. for
unused resource checks), but can filter out any warnings local to that
project. This is useful when you are using a library, but don't want
to see errors from that library which may not be under your control.

The way this is handled from the command line is that lint will only
report errors for projects you've referenced; e.g. if you run "lint
/foo/bar" this will show errors in /foo/bar, but exclude errors found
in the library project /foo/bar/../library".

However, there were several lint checks which needed additional fixes
for this, because (like the unused error detector) they gather data
from multiple projects and process and report in the after-project
hook.

In addition, inside Eclipse, the UI would always automatically include
libraries. This is sometimes what you want, and sometimes not what you
want, so this CL adds a new toggle menu item to the lint action menu,
"Skip Library Project Dependencies", off by default, which you can
select to make lint runs from eclipse include or exclude library
projects in the report.

Change-Id: Idf3167e818931525e0dd7661f5cdf3a3e69b6522