Don't warn about missing classes before a build
The missing class detector warns about classes declared in
the manifest, but not found on disk. In order not to warn
before you've built the project, it only does this if it has
seen at least one class.

However, it turned out that it would also look at library jars,
such that a not-built project using the support library would
be considered to have classes, and would therefore flag newly
added activities.

Change-Id: Iaf01fe3e3626bdb31d535c00aea584793d6f72b9