Check for unsupported Java packages
If you include a library in your class path which was not built as an
Android project it's possible for it to include references to classes
in the java.* and javax.* namespace which are not valid in Android.

This changeset adds a new lint detector for this. It augments the API
database used by the API Checker to record the set of platform
packages, and to be able to quickly query whether an internal class
reference is in one of the valid packages.

The new lint check runs only on libraries (since tools will typically
set up the right classpath to give immediate errors if you try to
access invalid packages from your own code).

This was requested in issue 39109.

Change-Id: Id1ce7982e683bae9a484e7b75d7e77a256ca4414