Improve double click handling in the XML editors
When you double click on the name portion of a resource url, it
selects just the name. Similarly when you double click on just
the class portion of a fully qualified class name, it selects
just the class. Finally, when you click on an attribute name
consisting of a namespace prefix and a local name, it selects
only the local name if you click within the local name region.

Change-Id: I050eb1493e570fd94719679e3eef2a40c92db2f8