Add resource renaming support
This changeset adds support for renaming resources.

There are several new hooks for initiating a resource rename:

(1) You can use the same keybinding as in Java files from XML files to
    initiate refactoring; for example, place the caret somewhere in
    @+id/foo or @string/bar and hit the refactoring keybinding, and a
    rename resource refactoring dialog shows up.

(2) Invoking Quick Assistant in an XML file (Ctrl/Cmd 1) will offer to
    rename the resource, if the caret is over a resource name.

(3) Renaming an XML or bitmap resource file, such as
    res/drawable-hdpi/foo.png, will now initiate the same XML resource
    naming machinery as above to update all resource references, plus
    it will also update all the other versions of the same file
    (e.g. in res/drawable-mdpi, res/drawable-xhdpi, etc.). Assuming an
    R field exists (e.g. the project has been built), it will also
    optionally update all Java field references.

(4) Invoking renaming in the layout editor (via the rename keybinding,
    or via the context menu, or via the property sheet's "..." button)
    will also initiate id resource refactoring. Editing the id
    directly in the inline editor for the id will pop up a dialog
    asking whether to update references as well, along with a "Do not
    ask again" checkbox.

(5) Finally, there is a renaming participant registered which will
    discover whether an R field is renamed, so if you go and rename
    R.layout.foo from Java, this will also kick in all of the above
    machinery - renaming layout files, updating resource references,
    etc.

If the renamed resource is in a library project, the refactoring
will also look at all the including projects and offer to update
references there as well.

Finally, this CL goes and fixes a few bugs in the existing refactoring
operations; in particular, making sure that they not only look at
files in layout/ but in all folder configurations containing layout
files. It also adds refactoring unit tests.

Change-Id: Ie88511a571b414fdc5be048e781fe29a34063cbf