Add AdtPlugin.getShell()
A lot of SWT and JFace methods require a Shell, and we had a lot of
the following calls sprinkled throughout the codebase:

AdtPlugin.getDisplay().getActiveShell()

However, getActiveShell() can return null in quite a few cases (where
it's not necessary), for example during focus transfers.

This CL adds a new method, AdtPlugin.getShell(), which first tries the
above call, but if that fails, will try harder to find another valid
shell (e.g. the first shell in Display.getShells()).

Change-Id: I3214ad56042be6eaf9b0fe0843c820e973fba8c0