Add sdktestutils package and consolidate some test utility code
This CL creates a new project, sdk/testutils, which is added to the
test projects for ADT and lint. It contains various test utility code,
some from ADT tests, some from lint tests, and some duplicated in
both, and some new utility code such as a better diff-er.

The CL also updates a bunch of tests; many golden files are updated to
the new test file diff output format, and some golden files were out
of date, and in a couple of cases I fixed some bugs I noticed in the
output.

Change-Id: I98c49c929916ae9bad24db45602e10ded6c526da