TaskHelper: properly parse Pkg.Revision property.
SDK Bug: 36987

Bug was that TaskHelper would fail when parsing a pk
revision with an RC value. A quick fix was done in CL
6d266a1bf5992d27fff4400c5d2bd30681c65c3d a while ago.

Revisit the fix by using the new FullRevision object
to parse the value and create a DeweyDecimal; should
avoid us trouble if we later change the format.

Change-Id: I5ad1449ebd2606255c27485e60b714b5cf83a308