SDK Manager: auto-select Support Library at startup.
When the SDK Manager starts, it auto-selects the
Windows driver on Windows. This change makes it
also auto-select the Support library extra package.
The rationale is that a lot of code now depends on it
(ADT templates, docs' sample code) so sooner or later
the user will need to install it anyway. We can help
newcomers by selecting it upfront.

Change-Id: I292c2d324bf68fe6b60f901fa5ecde0176ea5ec1