Make inefficient weight detector identify more candidates
The inefficient weight detector looks for horizontal layouts with
weights where all the children are layouts. It didn't have a very good
heuristic for finding out whether all children are layouts; it just
looked for whether the tag name contains "Layout".

This changeset generalizes this a bit, adding other layouts, as well
as <fragment> tags to the lint check.

It also updates the Master/Detail template to set the baseline
alignment attribute to false (to avoid triggering the updated lint
check), as well as android:textIsSelectable to clear another recent
new lint rule.

Change-Id: I887962b5e29a9ad0c5b5c01970b0c66f884125d8