Fix Rename Application Package such that it doesn't touch bin/
Fix the rename application package refactoring such that it only
modifies AndroidManifest.xml, not for example a copy in bin/.

Also added a diagnostic information message to the plain Java
package refactoring, shown when refactoring the application
package, explaining that the refactoring will not actually change
the manifest package, and explaining how to invoke the dedicated
refactoring to change it.

Also add unit tests for this refactoring.

Change-Id: I13054e033b9ee44210ca341f00c633d73a4418b4