Add flag to disable dex merging.
with dex.disable.merger=true in the project.properties,
it is now possible to revert to the old build mechanism where
every jar files is always passed to dex during dex'ing instead
of pre-dexing them only when they change.

This is to go around some issue with dexmerger that are only
triggered in some cases.

Change-Id: I7288ae4e1ae02b088c2f74f2f63c819c6587016a