Move lint unit tests into lint/cli
The unit tests for lint have lived in the lint/libs/lint_checks
library, but that's not a natural place since the cli library depends
on the lint_checks library, yet the lint_check *tests* also depend on
cli. Furthermore, the tests also test stuff in the cli library. Thus,
move the test sources over to the cli library, which also pulls in the
lint_checks library.

Change-Id: Ia8527ea894b287915ad12c04c16daf80ace67bdc