gltrace: Refresh tree view in a separate task.
When the currently displayed frame is changed, we immediately
refresh the tree displaying the list of calls, and the frame summary
view displaying the frame buffer as part of the frame change event
handler.

This CL moves these two operations to a separate task that is
scheduled after a short delay (250 ms). This allows us to combine
multiple frame change events and respond just once for all of them
together.

Change-Id: I4edade1e55cbd7394cba03d97b2ee61ca3a258b6