systrace: Support newer atrace options.
The atrace executable on the device now provides a list of
supported categories with the command atrace --list_categories.
This CL first tries to determine which version of atrace is
present on the device by running the list categories command
and attempting to parse its output.

For the older version, existing static "Capture Options" dialog
is displayed. For the newer version, only the supported categories
are displayed.

Change-Id: I300e7e652a9ae99f47d61a1669604802b74054a4